requirements.txt
asgiref==3.8.1
cryptography==45.0.4
dj-database-url==3.0.1
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
gunicorn==23.0.0
psycopg2-binary==2.9.11
python-decouple==3.8
sqlparse==0.5.3
